+++ src/lib/index.ts
@@ -1,4 +1,5 @@
 export { ELogLevel, setVerbosity } from './helpers/logger';
 export type { IHeightmap, IHeightmapCoords, IHeightmapSample } from './terrain/heightmap/i-heightmap';
 export { EPatchComputingMode, Terrain, type TerrainOptions, type TerrainStatistics } from './terrain/terrain';
-export type { ILocalMapData, IVoxelMap, IVoxelMaterial } from './terrain/voxelmap/i-voxel-map';
+export type { ILocalMapData, IVoxelMap, IVoxelMaterial } from './terrain/voxelmap/i-voxelmap';
+
